+++ laravel/resources/views/livewire/pages/course.blade.php
