+++ laravel/resources/views/pages/course/show.blade.php
@@ -1,8 +1,6 @@
 <x-app-layout>
     <div style="height: 100vh;">
-        
-        <livewire:course :courseId="$courseId" />
+        <livewire:pages.course :courseId="$courseId" />
         @livewireScripts
-    
     </div>
 </x-app-layout>
